Started work of porting the Sphinx API to a command line interface (CLI).
The CLI API has all of the features needed for BibleForge, but not being able to start a search at a specified id slows it down tremendously after a few loops.

Reverted sphixapi.phStarted work of porting the Sphinx API to a command line interface (CLI).

The CLI API has all of the features needed for BibleForge, but not being able to start a search at a specified id slows it down tremendously after a few loops.

Reverted sphixapi.php to be more like the original format due to regression.