Code cleanup and documentation.
Combined two of the replace() functions in prepare_search() for speed benefits.
Could also use replace(/\s*(?:-|\bNOT)\s/g, " -"), but it seems to be a little slower,
and it would create a leading space if the first word had a hyphen (-) in front of
it; however, the trim() function at the end would remove it.