Restructed the code to move most of the scrolling variables and all of functions into a closure.
An object with two methods is returned for the rest of the code to access.
The variables scroll_maxed_bottom, scroll_maxed_top, and scroll_pos should
also probalby be in the closure too, but at the moment, other functions
rely too heavily upon them.

More work is needed in this area.