Added notes about possibly checking to see if the cursor needs to be unhidden.
It would be nice not to call show_cursor() every time the mouse moves,
but that would either require accessing the DOM (via object.style.cursor)
or storing a value in a local variable.
The latter would be faster (though not necessarily noticeably),
but it would add several more lines of code, which would require more bandwidth and parsing.