Added new, beautiful input box.
Added a beautiful input box using CSS.
It doesn't look special in IE 8-, no surprise there, but it is functional.
Opera has minor glitches: the shadow background extends past the rounded corners, and the input box is too close to the top of the page.
Removed "BibleForge" text so that the input box will be centered correctly.
Should add a logo in the background (when a logo exists).

(NOTE: Chrome started acting strange.  It seems to be scrolling down after
loading both top and bottom results and the user clicks off of the search box.)