Added tooltip to query box for some explanation, and now it brings focus to the query box on start up.
It could be better to have some light text in the box, but that causes problems.
One problem is that there is hover CSS that darkens the text,
so that would have to be worked around.
Furthermore, if the query box is selected on start up, then that text would
not be visible anyway, at first.

Below is some code that would provide the basic features, but it has problems:
like the fact that onmouseout should really be onmouseleave (but only IE supports that).


q_obj.onblur = function ()
{
    if (this.value == "") {
        this.value = 'Keyword or Reference: "God so loved" or Romans 3:23';
    }
};

q_obj.onblur();

q_obj.onfocus = function ()
{
    if (this.value == 'Keyword or Reference: "God so loved" or Romans 3:23') {
        this.value = "";
    }
};

q_obj.onmouseover = q_obj.onfocus;
q_obj.onmouseout  = q_obj.onblur;