Begins the creation of a wrench context menu.
Currently, it uses transitionEnd (and it's variants) to provide a callback,
but this is not supported by many browsers (Firefox 3.6, IE 9(?)), and it
is difficult to create a graceful fallback because it is difficult to know
if transitionEnd is supported by the browser.

Also, fading out the menu may not be the most desirable action for usability's
sake since it adds a delay.  Windows doesn't fade out the menu.