Created (but not uses) the menu_maker() function to create menus.
This is mainly just for reference.  At this point, it doesn't seem worth
the effort of modifying the code to use it because it would make the code
less manageable, in my opinion.  The simplicity of the function illustrates
the small amount of code duplication if the functions are kept separate.