Replaced the gradient image in the query box with a nicer looking inset shadow.
Currently, this does not work in (most?) WebKit based browsers.  See Chromium
issue 29427 for more details:
http://code.google.com/p/chromium/issues/detail?id=29427.

On the other hand, it does work in Opera now.

Ironically, the base64 encoded GIF file is actually smaller than the CSS
for the inset shadow (because of all of the different prefixes).  But the
inset shadow also looks nicer because if follows the curves.