Integrated prepare_highlighter() into the language specific file.  (Still broken)
Now it returns an array as well.

filter_terms_for_highlighter() has very little language specific code and
should be moved out if possible.  It also might be best to move
prepare_highlighter() back out as well (but it would have to be largely
rewritten since it relies heavily on language specific code).