Change create_simple_ajax() into a constructor function called Create_easy_ajax().
I'm not sure if there is any benefit to doing this, but it seems to be the
typical way things are done.  The only difference I know of is that when
using the "new" operator, the "this" value of the function becomes a blank
object and lets you modify it by modifying "this."  I'm not sure if that
is beneficial, though.

Also, changed case because constructors begin with a capital letter by
convention.

The name change from "simple" to "easy" is meant to reflect its simple to
use characteristics, in contrast to being simple, i.e., limited.