Some documentation cleanup.
Merry Christmas!  May the holy child Jesus dwell in your hearts.

"And she shall bring forth a son, and thou shalt call his name JESUS:
for he shall save his people from their sins."
—Matthew 1:21

"And we know that the Son of God is come, and hath given us an
understanding, that we may know him that is true, and we are in him that is
true, even in his Son Jesus Christ. This is the true God, and eternal life."
—1 John 5:20