Adjusted JSLint options.
JSLint's indent options works now, but now the white option is broken.
Also turned on the "use strict" option for good measure.  The continue
option was just added too, and for now, that is turned on.  It would be
nice to be able to turn that off in the future.